src/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
psycopg2-binary
psutil
pydantic
bcrypt
aiofiles
itsdangerous
slowapi